backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg[binary]==3.2.3
pydantic==2.8.2
python-multipart==0.0.9
PyJWT==2.9.0
passlib==1.7.4
python-dotenv==1.0.1
email-validator==2.2.0
minio==7.2.7
python-multipart==0.0.9
